requirements.txt
requests
beautifulsoup4
PyQt6
arxiv
semanticscholar
accelerate
ddgs
google-generativeai
python-dotenv
pdfplumber
